Solr.Class

Description

This module defines the finally tagless Solr DSL. This style admits multiple interpreters, one of which (lazy ByteStrings) is provided by this library in the Solr.Query module.

Users should instead import either Solr.Query or Solr.Qualified.Query.

Solr language

class SolrExprSYM expr where

Solr expression.

Methods

num :: Float -> expr TNum

A num expression.

Example:

-- foo:5
query :: SolrQuery 'False 'False
query = "foo" =: num 5

true :: expr TBool

A true expression.

Example:

-- foo:true
query :: SolrQuery 'False 'False
query = "foo" =: true

false :: expr TBool

A false expression.

Example:

-- foo:false
query :: SolrQuery 'False 'False
query = "foo" =: false

word :: Text -> expr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for SolrExpr TWord, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:bar
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"

-- foo:bar
query :: SolrQuery 'False 'False
query = "foo" =: ("bar" :: SolrExpr TWord)

wild :: Text -> expr TWild

A single word that may contain wildcard characters ('?' and '*'), although the meaning of consecutive '*'s is probably ill-defined. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

Example:

-- foo:b?r
query :: SolrQuery 'False 'False
query = "foo" =: wild "b?r"

regex :: Text -> expr TRegex

A regular expression, whose syntax is described by http://lucene.apache.org/core/5_5_0/core/org/apache/lucene/util/automaton/RegExp.html?is-external=true.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

-- foo:/[mb]oat/
query :: SolrQuery 'False 'False
query = "foo" =: regex "[mb]oat"

phrase :: [expr TWord] -> expr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

Note that sometimes you may use the IsList instance for SolrExpr TPhrase, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:"bar baz"
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", "baz"] -- ok

-- foo:"bar b?z" (an invalid Solr query)
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", wild "b?z"] -- type error

-- foo:"bar b?z" (an invalid Solr query)
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", "b?z"] -- breaks word contract

Or, with OverloadedLists:

-- foo:"bar baz"
query :: SolrQuery 'False 'False
query = "foo" =: (["bar", "baz"] :: SolrExpr TPhrase)

(~:) :: FuzzableType a => expr a -> Int -> expr TFuzzed infix 6

The '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

Example:

-- foo:bar~1
query :: SolrQuery 'False 'False
query = "foo" =: word "bar" ~: 1

-- foo:"bar baz qux"~10
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", "baz", "qux"] ~: 10

to :: PrimType a => Boundary (expr a) -> Boundary (expr a) -> expr TRange infix 6

A range expression.

Example:

-- foo:[5 TO 10}
query :: SolrQuery 'False 'False
query = "foo" =: incl (num 5) `to` excl (num 10)

(^:) :: BoostableType a => expr a -> Float -> expr TBoosted infix 6

The '^' operator, which boosts its argument.

Example:

-- foo:bar^3.5
query :: SolrQuery 'False 'False
query = "foo" =: word "bar" ^: 3.5

-- foo:"bar baz"^3.5
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", "baz"] ^: 3.5

class SolrExprSYM expr => SolrQuerySYM expr query | query -> expr where

Solr query.

Associated Types

data ParamKey query :: * -> *

Different queries support different sets of local parameters. Each parameter is indexed by the type of its value.

Example:

data ParamKey MyQuery a where
  Foo :: ParamKey MyQuery Int
  Bar :: ParamKey MyQuery Bool
  Baz :: ParamKey MyQuery String

Methods

defaultField :: expr a -> query False False

A default field query.

Example:

-- foo
query :: SolrQuery 'False 'False
query = defaultField (word "foo")

(=:) :: Text -> expr a -> query False False infix 5

A field query.

Example:

-- foo:bar
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"

(&&:) :: query False False -> query False False -> query False False infixr 3

An AND query.

Example:

-- foo:bar AND baz:qux
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"
    &&: "baz" =: word "qux"

(||:) :: query False False -> query False False -> query False False infixr 2

An OR query.

Example:

-- foo:bar OR baz:qux
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"
    ||: "baz" =: word "qux"

(-:) :: query False False -> query False False -> query False False infixr 1

A NOT, '!', or '-' query.

Example:

-- foo:bar NOT baz:qux
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"
     -: "baz" =: word "qux"

(^=:) :: query False False -> Float -> query False False infixr 4

The '^=' constant score operator.

This is given right-fixity to reject queries like q ^= 1 ^= 2, which may very well be a valid Solr query (I haven't tested), but are nonetheless nonsense.

Example:

-- (foo:bar)^=3.5
query :: SolrQuery 'False 'False
query = "foo" =: word "bar" ^=: 3.5

neg :: query False False -> query True False

Negate a top-level query.

Only a top level query can be negated up to one time, hence the boolean tag that tracks whether or not a query has been negated.

Example:

-- -(foo:bar)
query :: SolrQuery 'True 'False
query = neg ("foo" =: word "bar")

params :: [Param query] -> query isNeg False -> query isNeg True

Add local parameters to a query.

Only a top level query can have up to one set of local parameters, hence the boolean tag that tracks whether or not they've been added.

Example:

-- {!df=foo}bar
query :: SolrQuery 'False 'True
query = params [paramDefaultField .= "foo"] (defaultField (word "bar"))

Derived combinators

fuzzy :: SolrExprSYM expr => expr TWord -> expr TFuzzed

Short-hand for fuzzing a word by 2. This is the default behavior of a Solr '~' operator without an integer added.

fuzzy e = e ~: 2

Example:

-- foo:bar~
query :: SolrQuery 'False 'False
query = "foo" =: fuzzy "bar"

gt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than range query.

gt e = excl e `to` star

Example:

-- foo:>5
-- foo:{5 TO *]
query :: SolrQuery 'False 'False
query = "foo" =: gt (num 5)

gte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than-or-equal-to range query.

gte e = incl e `to` star

Example:

-- foo:>=5
-- foo:[5 TO *]
query :: SolrQuery 'False 'False
query = "foo" =: gte (num 5)

lt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than range query.

 lt e = star `to` excl e

Example:

-- foo:<5
-- foo:[* TO 5}
query :: SolrQuery 'False 'False
query = "foo" =: lt (num 5)

lte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than-or-equal-to range query.

lte e = star `to` incl e

Example:

-- foo:<=5
-- foo:[* TO 5]
query :: SolrQuery 'False 'False
query = "foo" =: lte (num 5)

Local parameters

data Param query where

A parameter is built from a key and a value, whose type depends on the key.

Constructors

Param :: ParamKey query val -> val -> Param query

(.=) :: ParamKey query val -> val -> Param query

Infix constructor for Param.

Range query helpers

data Boundary a

An inclusive or exclusive expression for use in a range query, built with either incl, excl, or star.

The constructors are exported for use in interpreters.

Constructors

Inclusive a

Exclusive a

Star

incl :: SolrExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as inclusive, for use in a range query.

excl :: SolrExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as exclusive, for use in a range query.

star :: SolrExprSYM expr => Boundary (expr a)

'*' operator, signifying the minimum or maximun bound of a range. A [* TO *] query will require a type annotation.