Solr.Expr.Class

Solr expression language

class ExprSYM expr where

Solr expression.

Minimal complete definition

num, true, false, word, wild, regex, phrase, (~:), to, (^:)

Methods

num :: Float -> expr TNum

A num expression.

Examples

>>> compile [] ("foo" =: num 5 :: Query Expr)
"q=foo:5.0"

true :: expr TBool

A true expression.

Examples

>>> compile [] ("foo" =: true :: Query Expr)
"q=foo:true"

false :: expr TBool

A false expression.

Examples

>>> compile [] ("foo" =: false :: Query Expr)
"q=foo:false"

word :: Text -> expr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for Expr TWord, but usually an explicit type signature will be required (at the interpretation site or earlier).

Examples

>>> compile [] ("foo" =: word "bar" :: Query Expr)
"q=foo:bar"

wild :: Text -> expr TWild

A single word that may contain wildcard characters ('?' and '*'), although the meaning of consecutive '*'s is probably ill-defined. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

Examples

>>> compile [] ("foo" =: wild "b?r" :: Query Expr)
"q=foo:b?r"

regex :: Text -> expr TRegex

A regular expression, whose syntax is described by http://lucene.apache.org/core/5_5_0/core/org/apache/lucene/util/automaton/RegExp.html?is-external=true.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

Examples

>>> compile [] ("foo" =: regex "[mb]oat" :: Query Expr)
"q=foo:/[mb]oat/"

phrase :: [expr TWord] -> expr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

Examples

>>> compile [] ("foo" =: phrase ["bar", "baz"] :: Query Expr)
"q=foo:\"bar baz\""

(~:) :: Fuzzable a => expr a -> Int -> expr (TFuzzed a) infix 6

The '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

Examples

>>> compile [] ("foo" =: word "bar" ~: 1 :: Query Expr)
"q=foo:bar~1"

>>> compile [] ("foo" =: phrase ["bar", "baz", "qux"] ~: 10 :: Query Expr)
"q=foo:\"bar baz qux\"~10"

to :: Rangeable a => Boundary (expr a) -> Boundary (expr a) -> expr (TRanged a) infix 6

A range expression.

Examples

>>> compile [] ("foo" =: incl (num 5) `to` excl (num 10) :: Query Expr)
"q=foo:[5.0 TO 10.0}"

>>> compile [] ("foo" =: excl (word "bar") `to` star :: Query Expr)
"q=foo:{bar TO *]"

- Note the explicit type signature required for [* TO *] queries >>> compile [] ("foo" =: star to (star :: Boundary (Expr 'TNum)) :: Query Expr) "q=foo:[* TO *]"

(^:) :: Boostable a => expr a -> Float -> expr (TBoosted a) infix 6

The '^' operator, which boosts its argument.

Examples

>>> compile [] ("foo" =: word "bar" ^: 3.5 :: Query Expr)
"q=foo:bar^3.5"

>>> compile [] ("foo" =: phrase ["bar", "baz"] ^: 3.5 :: Query Expr)
"q=foo:\"bar baz\"^3.5"

Derived combinators

fuzzy :: ExprSYM expr => expr TWord -> expr (TFuzzed TWord)

Short-hand for fuzzing a word by 2. This is the default behavior of a Solr '~' operator without an integer added.

fuzzy e = e ~: 2

Examples

>>> compile [] ("foo" =: fuzzy "bar" :: Query Expr)
"q=foo:bar~2"

gt :: (ExprSYM expr, Rangeable a) => expr a -> expr (TRanged a)

Short-hand for a greater-than range query.

gt e = excl e `to` star

Examples

>>> compile [] ("foo" =: gt (num 5) :: Query Expr)
"q=foo:{5.0 TO *]"

gte :: (ExprSYM expr, Rangeable a) => expr a -> expr (TRanged a)

Short-hand for a greater-than-or-equal-to range query.

gte e = incl e `to` star

Examples

>>> compile [] ("foo" =: gte (num 5) :: Query Expr)
"q=foo:[5.0 TO *]"

lt :: (ExprSYM expr, Rangeable a) => expr a -> expr (TRanged a)

Short-hand for a less-than range query.

 lt e = star `to` excl e

Examples

>>> compile [] ("foo" =: lt (num 5) :: Query Expr)
"q=foo:[* TO 5.0}"

lte :: (ExprSYM expr, Rangeable a) => expr a -> expr (TRanged a)

Short-hand for a less-than-or-equal-to range query.

lte e = star `to` incl e

Examples

>>> compile [] ("foo" =: lte (num 5) :: Query Expr)
"q=foo:[* TO 5.0]"

Range expression helpers

data Boundary a

An inclusive or exclusive expression for use in a range query, built with either incl, excl, or star.

The constructors are exported for use in interpreters.

Constructors

Inclusive a

Exclusive a

Star

incl :: ExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as inclusive, for use in a range query.

excl :: ExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as exclusive, for use in a range query.

star :: ExprSYM expr => Boundary (expr a)

'*' operator, signifying the minimum or maximun bound of a range. A [* TO *] query will require a type annotation.

Named operators

fuzz :: (ExprSYM expr, Fuzzable a) => expr a -> Int -> expr (TFuzzed a)

Named version of (~:).

boost :: (ExprSYM expr, Boostable a) => expr a -> Float -> expr (TBoosted a)

Named version of (^:).