Solr.Expr.Initial.Typed

Expression type

data Expr :: SolrType -> * where

A typed, initially-encoded Solr expression.

Constructors

ENum :: Float -> Expr TNum

ETrue :: Expr TBool

EFalse :: Expr TBool

EWord :: Text -> Expr TWord

EWild :: Text -> Expr TWild

ERegex :: Text -> Expr TRegex

EPhrase :: [Expr TWord] -> Expr TPhrase

EFuzz :: Fuzzable a => Expr a -> Int -> Expr (TFuzzed a)

ETo :: Rangeable a => Boundary (Expr a) -> Boundary (Expr a) -> Expr (TRanged a)

EBoost :: Boostable a => Expr a -> Float -> Expr (TBoosted a)

typeCheck :: Expr a -> (forall ty. Maybe (Expr ty) -> r) -> r

Type check an untyped Solr expression. Note the untyped Expr on the way in is not the same as the typed Expr on the way out.

typeCheck u k =
  case typeCheck' u of
    Nothing -> k Nothing
    Just (SomeExpr e) -> k (Just e)

reinterpret :: ExprSYM expr => Expr ty -> expr ty

Reinterpret a Solr expression.

Re-exports

module Solr.Expr.Class