Solr.Internal.Class.Query

Description

This module defines the finally tagless Solr DSL. This style admits multiple interpreters, two of which (lazy ByteStrings and an initial encoding) are provided by this library, in the Solr.Query and Solr.Query.Initial modules, respectively.

Ordinary users should not normally have to import this module.

Solr query language

class SolrExprSYM expr => SolrQuerySYM expr query where

Solr query language.

Associated Types

data ParamKey query :: * -> *

Different queries support different sets of local parameters. Each parameter is indexed by the type of its value.

Example:

data ParamKey MyQuery a where
  Foo :: ParamKey MyQuery Int
  Bar :: ParamKey MyQuery Bool
  Baz :: ParamKey MyQuery String

Methods

defaultField :: expr a -> query expr

A default field query.

>>> defaultField (word "foo") :: SolrQuery SolrExpr
q=foo

(=:) :: Text -> expr a -> query expr infix 5

A field query.

>>> "foo" =: word "bar" :: SolrQuery SolrExpr
q=foo:bar

(&&:) :: query expr -> query expr -> query expr infixr 3

An AND query.

>>> "foo" =: word "bar" &&: "baz" =: word "qux" :: SolrQuery SolrExpr
q=(foo:bar AND baz:qux)

(||:) :: query expr -> query expr -> query expr infixr 2

An OR query.

>>> "foo" =: word "bar" ||: "baz" =: word "qux" :: SolrQuery SolrExpr
q=(foo:bar OR baz:qux)

(-:) :: query expr -> query expr -> query expr infixr 1

A NOT, '!', or '-' query.

>>> "foo" =: word "bar" -: "baz" =: word "qux" :: SolrQuery SolrExpr
q=(foo:bar NOT baz:qux)

(^=:) :: query expr -> Float -> query expr infixr 4

The '^=' constant score operator.

This is given right-fixity to reject queries like q ^= 1 ^= 2.

>>> "foo" =: word "bar" ^=: 3.5 :: SolrQuery SolrExpr
q=foo:bar^=3.5

neg :: query expr -> query expr

Negate a query.

>>> neg ("foo" =: word "bar") :: SolrQuery SolrExpr
q=-foo:bar

params :: [Param query] -> query expr -> query expr

Add local parameters to a query.

>>> params [paramDefaultField .= "foo"] (defaultField (word "bar")) :: SolrQuery SolrExpr
q={!df=foo}bar

Local parameters

data Param query where

A parameter is built from a key and a value, whose type depends on the key.

Constructors

Param :: ParamKey query val -> val -> Param query

(.=) :: ParamKey query val -> val -> Param query

Infix constructor for Param.

Named operators

field :: SolrQuerySYM expr query => Text -> expr a -> query expr

Named version of (=:).

and :: SolrQuerySYM expr query => query expr -> query expr -> query expr infixr 3

Named version of (&&:).

or :: SolrQuerySYM expr query => query expr -> query expr -> query expr infixr 2

Named version of (||:).

not :: SolrQuerySYM expr query => query expr -> query expr -> query expr infixr 1

Named version of (-:).

score :: SolrQuerySYM expr query => query expr -> Float -> query expr infixr 4

Named version of (^=:).

Re-exports

module Solr.Internal.Class.Expr