Solr.Param

Documentation

class HasParamDefaultField query where

The class of queries that support the 'df' local parameter.

Methods

paramDefaultField :: ParamKey query Text

The 'df' local parameter.

Example:

-- {!df=foo}bar
query :: SolrQuery
query = params [paramDefaultField .= "foo"] (defaultField (word "bar"))

class HasParamOp query where

The class of queries that support the 'op' local parameter.

Methods

paramOp :: ParamKey query Text

The 'op' local parameter.

Stringly typed to avoid a clunky sum type like

data Val = And | Or | ...

which seems to have little value in cases like this. Instead, just pass "AND", "OR", ...

Example:

-- {!q.op=AND}foo bar
query :: SolrQuery
query = params [paramOp .= "AND"] (defaultField (word "foo") <> defaultField (word "bar"))

class HasParamCache query where

The class of queries that support the 'cache' local parameter.

Methods

paramCache :: ParamKey query Bool

The 'cache' local parameter.

Example:

-- {!cache=false}foo:bar
query :: SolrFilterQuery
query = params [paramCache .= False] ("foo" =: word "bar")

class HasParamCost query where

The class of queries that support the 'cost' local parameter.

Methods

paramCost :: ParamKey query Int

The 'cost' local parameter.

Example:

-- {!cost=5}foo:bar
query :: SolrFilterQuery
query = params [paramCost .= 5] ("foo" =: word "bar")