Solr.Qualified.Class

Description

This module is an alternative to Solr.Class that does not export any operators, and is intended to be imported qualified, because it contains function names that clash with the Prelude.

import qualified Solr.Qualified.Class as Solr

Here is a quick conversion guide:

(~:)  = fuzz
(^:)  = boost
(=:)  = field
(&&:) = and
(||:) = or
(-:)  = not
(^=:) = score

Named operators

fuzz :: (SolrExprSYM expr, FuzzableType a) => expr a -> Int -> expr TFuzzed

Named version of (~:).

boost :: (SolrExprSYM expr, BoostableType a) => expr a -> Float -> expr TBoosted

Named version of (^:).

field :: SolrQuerySYM expr query => Text -> expr a -> query

Named version of (=:).

and :: SolrQuerySYM expr query => query -> query -> query infixr 3

Named version of (&&:).

or :: SolrQuerySYM expr query => query -> query -> query infixr 2

Named version of (||:).

not :: SolrQuerySYM expr query => query -> query -> query infixr 1

Named version of (-:).

score :: SolrQuerySYM expr query => query -> Float -> query infixr 4

Named version of (^=:).

Re-exports

module Solr.Class