Solr.Query

Query class

class (Coercible query Builder, Default (LocalParams query)) => Query query

Minimal complete definition

compileLocalParams

Associated Types

data LocalParams query :: *

should :: Query query => query -> query

This clause should occur.

must :: Query query => query -> query

This clause must occur.

mustNot :: Query query => query -> query

This clause must not occur.

filt :: Query query => query -> query

This clause must occur, but it does not participate in scoring.

data SomeQuery

SomeQuery is a simple wrapper around a Query that enables composition through its Monoid instance.

It has no LocalParams of its own - you can only create them with def.

someQuery :: Query query => LocalParams query -> query -> SomeQuery

Create a SomeQuery from a Query and its LocalParams.

Compiling a Query

compile :: [Param] -> Text

Compile a list of Param to a lazy Text.

Query params

data Param

A query parameter.

fl :: Text -> Param

The 'fl' query parameter.

fq :: Query query => FilterParams query -> query -> Param

The 'fq' query parameter.

q :: Query query => LocalParams query -> query -> Param

The 'q' query parameter.

rows :: Int -> Param

The 'rows' query parameter.

sortAsc :: Text -> Param

The 'sort' query parameter (ascending).

sortDesc :: Text -> Param

The 'sort' query parameter (descending).

start :: Int -> Param

The 'start' query parameter.

Re-exports

(&) :: a -> (a -> b) -> b infixl 1

& is a reverse application operator. This provides notational convenience. Its precedence is one higher than that of the forward application operator $, which allows & to be nested in $.

Since: 4.8.0.0

def :: Default a => a

The default value for this type.