Solr.Query.Class

Description

This module defines the finally tagless Solr DSL. This style admits multiple interpreters, two of which (lazy Texts and an initial encoding) are provided by this library, in the Solr.Query and Solr.Query.Initial modules, respectively.

Ordinary users should not normally have to import this module.

Solr query language

class ExprSYM expr => QuerySYM expr query where

Solr query language.

Minimal complete definition

defaultField, (=:), (&&:), (||:), (-:), (^=:), neg

Methods

defaultField :: expr a -> query expr

A default field query.

Examples

>>> compile [] (defaultField (word "foo") :: Query Expr)
"q=foo"

(=:) :: Text -> expr a -> query expr infix 5

A field query.

Examples

>>> compile [] ("foo" =: word "bar" :: Query Expr)
"q=foo:bar"

(&&:) :: query expr -> query expr -> query expr infixr 3

An AND query.

Examples

>>> compile [] ("foo" =: word "bar" &&: "baz" =: word "qux" :: Query Expr)
"q=(foo:bar AND baz:qux)"

(||:) :: query expr -> query expr -> query expr infixr 2

An OR query.

Examples

>>> compile [] ("foo" =: word "bar" ||: "baz" =: word "qux" :: Query Expr)
"q=(foo:bar OR baz:qux)"

(-:) :: query expr -> query expr -> query expr infixr 1

A NOT, '!', or '-' query.

Examples

>>> compile [] ("foo" =: word "bar" -: "baz" =: word "qux" :: Query Expr)
"q=(foo:bar NOT baz:qux)"

(^=:) :: query expr -> Float -> query expr infixr 4

The '^=' constant score operator.

This is given right-fixity to reject queries like q ^= 1 ^= 2.

Examples

>>> compile [] ("foo" =: word "bar" ^=: 3.5 :: Query Expr)
"q=foo:bar^=3.5"

neg :: query expr -> query expr

Negate a query.

Examples

>>> compile [] (neg ("foo" =: word "bar") :: Query Expr)
"q=-foo:bar"

Named operators

field :: QuerySYM expr query => Text -> expr a -> query expr

Named version of (=:).

and :: QuerySYM expr query => query expr -> query expr -> query expr infixr 3

Named version of (&&:).

or :: QuerySYM expr query => query expr -> query expr -> query expr infixr 2

Named version of (||:).

not :: QuerySYM expr query => query expr -> query expr -> query expr infixr 1

Named version of (-:).

score :: QuerySYM expr query => query expr -> Float -> query expr infixr 4

Named version of (^=:).

Re-exports

module Solr.Expr.Class