Solr.Query.Initial

Description

An initial encoding of a Solr query. This is an alternative interpretation of the Solr language that is more amenable to parsing from arbitrary user input and applying query transformations.

Initially-encoded Query

data Query expr

An initial encoding of a Solr query.

Constructors

QDefaultField (expr a)

QField Text (expr a)

QAnd (Query expr) (Query expr)

QOr (Query expr) (Query expr)

QNot (Query expr) (Query expr)

QScore (Query expr) Float

QNeg (Query expr)

QAppend (Query expr) (Query expr)

typeCheck :: Query Expr -> Maybe (Query Expr)

Type check an untyped Solr query. Note the untyped Expr on the way in is not the same as the typed Expr on the way out.

factor :: Query expr -> Query expr

Factor a Solr query into a canonical form (e.g. perform double-negation elimination). Check the source code for all transformations performed.

reinterpret :: forall expr query. (QuerySYM expr query, Semigroup (query expr)) => Query Expr -> query expr

Reinterpret an initially-encoded Query to some other interpretation.

This may be useful for reinterpreting a Query as a lazy Text after it's been type checked and factored with the machinery in this module.

Re-exports

module Solr.Param

module Solr.Query.Class