Solr.Query.Lucene

Lucene query

data LuceneQuery

defaultField :: LuceneExpr ty -> LuceneQuery

A default field query.

(=:) :: Text -> LuceneExpr ty -> LuceneQuery infix 7

A field query.

field :: Text -> LuceneExpr ty -> LuceneQuery

Named version of (=:).

(^=:) :: LuceneQuery -> Float -> LuceneQuery infixr 6

The '^=' constant score operator.

This is given right-fixity to reject queries like q ^= 1 ^= 2.

score :: LuceneQuery -> Float -> LuceneQuery infixr 6

Named version of (^=:).

Lucene expression

data LuceneExpr t

A lucene expression.

int expression

int :: Int64 -> LuceneExpr TNum

An int expression.

float expression

float :: Double -> LuceneExpr TNum

A float expression.

bool expressions

true :: LuceneExpr TBool

A true expression.

false :: LuceneExpr TBool

A false expression.

word expression

word :: Text -> LuceneExpr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for LuceneExpr TWord, but usually an explicit type signature will be required (at the interpretation site or earlier).

wild expression

wild :: Text -> LuceneExpr TWild

A single word that may contain wildcard characters ('?' and '*'), although the meaning of consecutive '*'s is probably ill-defined. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

regex expression

regex :: Text -> LuceneExpr TRegex

A regular expression.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

phrase expression

phrase :: [LuceneExpr TWord] -> LuceneExpr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

datetime expression

datetime :: IsDateTime a => a -> LuceneExpr TDateTime

A DateTime expression. This may either be a timestamp (UTCTime), or a "truncated" DateTime such as (2015, 5, 12).

data DateTime

DateTime literals. DateTime expressions are constructed using the internal IsDateTime typeclass, for which there exist the following instances:

instance IsDateTime UTCTime
instance IsDateTime Year
instance IsDateTime (Year, Month)
instance IsDateTime (Year, Month, Day)
instance IsDateTime (Year, Month, Day, Hour)
instance IsDateTime (Year, Month, Day, Hour, Minute)
instance IsDateTime (Year, Month, Day, Hour, Minute, Second)
instance IsDateTime (Year, Month, Day, Hour, Minute, Second, Millisecond)

type Year = Int

Year.

type Month = Int

1-indexed month. Clamped to the range 1-12.

type Day = Int

1-indexed day. Clamped to the range 1-31.

type Hour = Int

Hour. Clamped to the range 0-23.

type Minute = Int

Minute. Clamped to the range 0-59.

type Second = Int

Second. Clamped to the range 0-60.

type Millisecond = Double

Millisecond. Clamped to the range 0-99.999.

fuzz expression

(~:) :: Fuzzable a => LuceneExpr a -> Int -> LuceneExpr TFuzzy infix 9

The '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

fuzz :: Fuzzable a => LuceneExpr a -> Int -> LuceneExpr TFuzzy

Named version of (~:).

fuzzy :: LuceneExpr TWord -> LuceneExpr TFuzzy

Short-hand for fuzzing a word by 2. This is the default behavior of a Solr '~' operator without an integer added.

fuzzy e = e ~: 2

boost expression

(^:) :: Boostable a => LuceneExpr a -> Float -> LuceneExpr TBoosted infix 9

The '^' operator, which boosts its argument.

boost :: Boostable a => LuceneExpr a -> Float -> LuceneExpr TBoosted

Named version of (^:).

range expressions

to :: Rangeable a b => Boundary a -> Boundary b -> LuceneExpr TRange infix 9

A range expression.

gt :: Rangeable a TAny => LuceneExpr a -> LuceneExpr TRange

Short-hand for a greater-than range query.

gt e = excl e `to` star

gte :: Rangeable a TAny => LuceneExpr a -> LuceneExpr TRange

Short-hand for a greater-than-or-equal-to range query.

gte e = incl e `to` star

lt :: Rangeable TAny a => LuceneExpr a -> LuceneExpr TRange

Short-hand for a less-than range query.

 lt e = star `to` excl e

lte :: Rangeable TAny a => LuceneExpr a -> LuceneExpr TRange

Short-hand for a less-than-or-equal-to range query.

lte e = star `to` incl e

data Boundary ty

An inclusive or exclusive expression for use in a range query, built with either incl, excl, or star.

The constructors are exported for use in interpreters.

incl :: LuceneExpr a -> Boundary a

Mark an expression as inclusive, for use in a range query.

excl :: LuceneExpr a -> Boundary a

Mark an expression as exclusive, for use in a range query.

star :: Boundary TAny

'*' operator, signifying the minimum or maximun bound of a range.

spatial predicate expressions

intersects :: Shape -> LuceneExpr TSpatialPredicate

'Intersects' spatial predicate.

isWithin :: Shape -> LuceneExpr TSpatialPredicate

'IsWithin' spatial predicate.

data Shape

A shape.

polygon :: [(Double, Double)] -> Shape

A POLYGON shape.

Lucene expression types

data LuceneExprTy

Constructors

TAny

TNum

TBool

TWord

TWild

TRegex

TPhrase

TDateTime

TFuzzy

TBoosted

TRange

TSpatialPredicate

class Fuzzable ty

words and phrases can fuzzed by the ~: operator.

class Boostable ty

words and phrases can be boosted by the ^: operator.

class Rangeable a b

ints, floats, words, and datetimes can to range expression.

Local parameters

df :: Text -> LocalParams LuceneQuery -> LocalParams LuceneQuery

The 'df' local parameter.

opAnd :: LocalParams LuceneQuery -> LocalParams LuceneQuery

The 'op=AND' local parameter.

opOr :: LocalParams LuceneQuery -> LocalParams LuceneQuery

The 'op=OR' local parameter.

Re-exports

data UTCTime :: *

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.