Solr.Query.Param

Description

Query parameters.

Documentation

class HasParamDefaultField query where

The class of queries that support the 'df' local parameter.

Methods

paramDefaultField :: ParamKey query Text

The 'df' local parameter.

>>> params [paramDefaultField .= "foo"] (defaultField (word "bar")) :: SolrQuery SolrExpr
q={!df=foo}bar

class HasParamOp query where

The class of queries that support the 'op' local parameter.

Methods

paramOp :: ParamKey query Text

The 'op' local parameter.

Stringly typed to avoid a clunky sum type like

data Operator = And | Or | ...

which seems to have little value in cases like this. Instead, just pass "AND", "OR", ...

>>> params [paramOp .= "AND"] (defaultField (word "foo") <> defaultField (word "bar")) :: SolrQuery SolrExpr
q={!q.op=AND}foo bar

class HasParamCache query where

The class of queries that support the 'cache' local parameter.

Methods

paramCache :: ParamKey query Bool

The 'cache' local parameter.

>>> params [paramCache .= False] ("foo" =: word "bar") :: SolrFilterQuery SolrExpr
fq={!cache=false}foo:bar

class HasParamCost query where

The class of queries that support the 'cost' local parameter.

Methods

paramCost :: ParamKey query Int

The 'cost' local parameter.

>>> params [paramCost .= 5] ("foo" =: word "bar") :: SolrFilterQuery SolrExpr
fq={!cost=5}foo:bar