Solr.Expr.Initial.Typed

Description

This module is not yet complete.

Expression type

data SolrExpr :: SolrType -> * where

A Solr expression.

Constructors

ENum :: Float -> SolrExpr TNum

ETrue :: SolrExpr TBool

EFalse :: SolrExpr TBool

EWord :: Text -> SolrExpr TWord

EWild :: Text -> SolrExpr TWild

ERegex :: Text -> SolrExpr TRegex

EPhrase :: [SolrExpr TWord] -> SolrExpr TPhrase

EFuzz :: FuzzableType a => SolrExpr a -> Int -> SolrExpr TFuzzed

ETo :: PrimType a => Boundary (SolrExpr a) -> Boundary (SolrExpr a) -> SolrExpr TRange

EBoost :: BoostableType a => SolrExpr a -> Float -> SolrExpr TBoosted

Expression construction

num :: SolrExprSYM expr => Float -> expr TNum

A num expression.

>>> "foo" =: num 5 :: SolrQuery SolrExpr
q=foo:5.0

true :: SolrExprSYM expr => expr TBool

A true expression.

>>> "foo" =: true :: SolrQuery SolrExpr
q=foo:true

false :: SolrExprSYM expr => expr TBool

A false expression.

>>> "foo" =: false :: SolrQuery SolrExpr
q=foo:false

word :: SolrExprSYM expr => Text -> expr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for SolrExpr TWord, but usually an explicit type signature will be required (at the interpretation site or earlier).

>>> "foo" =: word "bar" :: SolrQuery SolrExpr
q=foo:bar

wild :: SolrExprSYM expr => Text -> expr TWild

A single word that may contain wildcard characters ('?' and '*'), although the meaning of consecutive '*'s is probably ill-defined. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

>>> "foo" =: wild "b?r" :: SolrQuery SolrExpr
q=foo:b?r

regex :: SolrExprSYM expr => Text -> expr TRegex

A regular expression, whose syntax is described by http://lucene.apache.org/core/5_5_0/core/org/apache/lucene/util/automaton/RegExp.html?is-external=true.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

>>> "foo" =: regex "[mb]oat" :: SolrQuery SolrExpr
q=foo:/[mb]oat/

phrase :: SolrExprSYM expr => [expr TWord] -> expr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

>>> "foo" =: phrase ["bar", "baz"] :: SolrQuery SolrExpr
q=foo:"bar baz"

(~:) :: (SolrExprSYM expr, FuzzableType a) => expr a -> Int -> expr TFuzzed

The '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

>>> "foo" =: word "bar" ~: 1 :: SolrQuery SolrExpr
q=foo:bar~1

>>> "foo" =: phrase ["bar", "baz", "qux"] ~: 10 :: SolrQuery SolrExpr
q=foo:"bar baz qux"~10

fuzz :: (SolrExprSYM expr, FuzzableType a) => expr a -> Int -> expr TFuzzed

Named version of (~:).

fuzzy :: SolrExprSYM expr => expr TWord -> expr TFuzzed

Short-hand for fuzzing a word by 2. This is the default behavior of a Solr '~' operator without an integer added.

fuzzy e = e ~: 2

>>> "foo" =: fuzzy "bar" :: SolrQuery SolrExpr
q=foo:bar~2

to :: (SolrExprSYM expr, PrimType a) => Boundary (expr a) -> Boundary (expr a) -> expr TRange

A range expression.

>>> "foo" =: incl (num 5) `to` excl (num 10) :: SolrQuery SolrExpr
q=foo:[5.0 TO 10.0}

>>> "foo" =: excl (word "bar") `to` star :: SolrQuery SolrExpr
q=foo:{bar TO *]

- Note the explicit type signature required for [* TO *] queries >>> "foo" =: star to (star :: Boundary (SolrExpr 'TNum)) :: SolrQuery SolrExpr q=foo:[* TO *]

incl :: SolrExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as inclusive, for use in a range query.

excl :: SolrExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as exclusive, for use in a range query.

star :: SolrExprSYM expr => Boundary (expr a)

'*' operator, signifying the minimum or maximun bound of a range. A [* TO *] query will require a type annotation.

gt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than range query.

gt e = excl e `to` star

>>> "foo" =: gt (num 5) :: SolrQuery SolrExpr
q=foo:{5.0 TO *]

gte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than-or-equal-to range query.

gte e = incl e `to` star

>>> "foo" =: gte (num 5) :: SolrQuery SolrExpr
q=foo:[5.0 TO *]

lt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than range query.

 lt e = star `to` excl e

>>> "foo" =: lt (num 5) :: SolrQuery SolrExpr
q=foo:[* TO 5.0}

lte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than-or-equal-to range query.

lte e = star `to` incl e

>>> "foo" =: lte (num 5) :: SolrQuery SolrExpr
q=foo:[* TO 5.0]

(^:) :: (SolrExprSYM expr, BoostableType a) => expr a -> Float -> expr TBoosted

The '^' operator, which boosts its argument.

>>> "foo" =: word "bar" ^: 3.5 :: SolrQuery SolrExpr
q=foo:bar^3.5

>>> "foo" =: phrase ["bar", "baz"] ^: 3.5 :: SolrQuery SolrExpr
q=foo:"bar baz"^3.5

boost :: (SolrExprSYM expr, BoostableType a) => expr a -> Float -> expr TBoosted

Named version of (^:).

Type checking

typeCheckSolrExpr :: SolrExpr a -> r -> (forall ty. SolrExpr ty -> r) -> r

Typecheck an untyped Solr expression. Note the SolrExpr on the way in is not the same as the SolrExpr on the way out.