Solr.Expr.Initial.Typed

Expression type

data SolrExpr :: SolrType -> * where

A typed, initially-encoded Solr expression.

Constructors

ENum :: Float -> SolrExpr TNum

ETrue :: SolrExpr TBool

EFalse :: SolrExpr TBool

EWord :: Text -> SolrExpr TWord

EWild :: Text -> SolrExpr TWild

ERegex :: Text -> SolrExpr TRegex

EPhrase :: [SolrExpr TWord] -> SolrExpr TPhrase

EFuzz :: Fuzzable a => SolrExpr a -> Int -> SolrExpr (TFuzzed a)

ETo :: Rangeable a => Boundary (SolrExpr a) -> Boundary (SolrExpr a) -> SolrExpr (TRanged a)

EBoost :: Boostable a => SolrExpr a -> Float -> SolrExpr (TBoosted a)

data SomeSolrExpr

Existential wrapper around SolrExpr.

Constructors

forall ty . SomeSolrExpr (SolrExpr ty)

Type checking

typeCheckSolrExpr :: SolrExpr a -> (forall ty. Maybe (SolrExpr ty) -> r) -> r

Type check an untyped Solr expression. Note the untyped SolrExpr on the way in is not the same as the typed SolrExpr on the way out.

typeCheckSolrExpr u k =
  case typeCheckSolrExpr' u of
    Nothing -> k Nothing
    Just (SomeSolrExpr e) -> k (Just e)

typeCheckSolrExpr' :: SolrExpr a -> Maybe SomeSolrExpr

Like typeCheckSolrExpr, but return an existential type rather than use rank-2 continuation passing style, if you prefer.

Reinterpretation

reinterpretSolrExpr :: SolrExprSYM expr => SolrExpr ty -> expr ty

Reinterpret a Solr expression.

Re-exports

module Solr.Internal.Class.Expr