Solr.Qualified.Class

Description

This module is an alternative to Solr.Class that does not export any operators, and is intended to be imported qualified, because it contains function names that clash with the Prelude.

import qualified Solr.Qualified.Class as Solr

Here is a quick conversion guide:

(~:)  = fuzz
(^:)  = boost
(=:)  = field
(&&:) = and
(||:) = or
(-:)  = not
(^=:) = score

Solr language

class SolrExprSYM expr

Solr expression.

Minimal complete definition

int, true, false, word, wild, regex, phrase, (~:), to, (^:)

int :: SolrExprSYM expr => Int -> expr TInt

An int expression.

Note that sometimes you may use the Num instance for SolrExpr TInt, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:5
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.int 5)

true :: SolrExprSYM expr => expr TBool

A true expression.

Example:

-- foo:true
query :: Solr.SolrQuery 'False
query = Solr.field "foo" Solr.true

false :: SolrExprSYM expr => expr TBool

A false expression.

Example:

-- foo:false
query :: Solr.SolrQuery 'False
query = Solr.field "foo" Solr.false

word :: SolrExprSYM expr => Text -> expr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for SolrExpr TWord, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:bar
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.word "bar")

-- foo:bar
query :: Solr.SolrQuery 'False
query = Solr.field "foo" ("bar" :: Solr.SolrExpr TWord)

wild :: SolrExprSYM expr => Text -> expr TWild

A single word that may contain wildcard characters ('?' and '*'), although the meaning of consecutive '*'s is probably ill-defined. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

Example:

-- foo:b?r
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.wild "b?r")

regex :: SolrExprSYM expr => Text -> expr TRegex

A regular expression, whose syntax is described by http://lucene.apache.org/core/5_5_0/core/org/apache/lucene/util/automaton/RegExp.html?is-external=true.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

-- foo:/[mb]oat/
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.regex "[mb]oat")

phrase :: SolrExprSYM expr => [expr TWord] -> expr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

Note that sometimes you may use the IsList instance for SolrExpr TPhrase, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:"bar baz"
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.phrase ["bar", "baz"]) -- ok

-- foo:"bar b?z" (an invalid Solr query)
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.phrase ["bar", Solr.wild "b?z"]) -- type error

-- foo:"bar b?z" (an invalid Solr query)
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.phrase ["bar", "b?z"]) -- breaks word contract

Or, with OverloadedLists:

-- foo:"bar baz"
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (["bar", "baz"] :: Solr.SolrExpr TPhrase)

fuzz :: (SolrExprSYM expr, FuzzableType a) => expr a -> Int -> expr (TFuzzed a)

The '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

This will have one of the following two types:

fuzz :: SolrExpr TWord   -> Int -> SolrExpr TFuzzyWord   -- Int must be 0, 1, or 2
fuzz :: SolrExpr TPhrase -> Int -> SolrExpr TFuzzyPhrase -- Int must be positive

Example:

-- foo:bar~1
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.fuzz (Solr.word "bar") 1)

-- foo:"bar baz qux"~10
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.fuzz (Solr.phrase ["bar", "baz", "qux"]) 10)

to :: (SolrExprSYM expr, PrimType a) => Boundary (expr a) -> Boundary (expr a) -> expr TRange infix 6

A range expression.

This will have one of the following two types:

to :: Boundary (SolrExpr TWord) -> Boundary (SolrExpr TWord) -> SolrExpr TRange
to :: Boundary (SolrExpr TInt)  -> Boundary (SolrExpr TInt)  -> SolrExpr TRange

Example:

-- foo:[5 TO 10}
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.incl (Solr.int 5) `Solr.to` Solr.excl (Solr.int 10))

boost :: (SolrExprSYM expr, BoostableType a) => expr a -> Float -> expr (TBoosted a)

The '^' operator, which boosts its argument.

This will have one of the following two types:

boost :: SolrExpr TWord   -> Float -> SolrExpr TBoostedWord
boost :: SolrExpr TPhrase -> Float -> SolrExpr TBoostedPhrase

Example:

-- foo:bar^3.5
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.boost (Solr.word "bar") 3.5)

-- foo:"bar baz"^3.5
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.boost (Solr.phrase ["bar", "baz"]) 3.5)

class SolrExprSYM expr => SolrQuerySYM expr query | query -> expr

Solr query.

Minimal complete definition

defaultField, (=:), (&&:), (||:), (-:), (^=:), localParams

defaultField :: SolrQuerySYM expr query => expr a -> query False

A default field query.

Example:

-- foo
query :: Solr.SolrQuery 'False
query = Solr.defaultField (Solr.word "foo")

field :: SolrQuerySYM expr query => Text -> expr a -> query False

A field query.

Example:

-- foo:bar
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.word "bar")

and :: SolrQuerySYM expr query => query False -> query False -> query False infixr 3

An AND query.

Example:

-- foo:bar AND baz:qux
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.word "bar") `Solr.and` Solr.field "baz" (Solr.word "qux")

or :: SolrQuerySYM expr query => query False -> query False -> query False infixr 2

An OR query.

Example:

-- foo:bar OR baz:qux
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.word "bar") `Solr.or` Solr.field "baz" (Solr.word "qux")

not :: SolrQuerySYM expr query => query False -> query False -> query False infixr 1

A NOT, '!' or '-' query.

Example:

-- foo:bar NOT baz:qux
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.word "bar") `Solr.not` Solr.field "baz" (Solr.word "qux")

score :: SolrQuerySYM expr query => query False -> Float -> query False infixr 4

The '^=' constant score operator.

This is given right-fixity to reject queries like q ^= 1 ^= 2, which may very well be a valid Solr query (I haven't tested), but are nonetheless nonsense.

Example:

-- (foo:bar)^=3.5
query :: Solr.SolrQuery 'False
query = Solr.score (Solr.field "foo" (Solr.word "bar")) 3.5

localParams :: SolrQuerySYM expr query => LocalParams query -> query False -> query True

Add local parameters to a query. A query can only have one set of local parameters, hence the boolean tag that tracks whether or not they've been added.

Example:

-- {!df=foo}bar
query :: Solr.SolrQuery 'True
query = Solr.localParams (Solr.paramDefaultField "foo") (Solr.defaultField (Solr.word "bar"))

Derived combinators

fuzzy :: SolrExprSYM expr => expr TWord -> expr TFuzzyWord

Short-hand for fuzzing a word by 2. This is the default behavior of a Solr '~' operator without an integer added.

fuzzy e = fuzz e 2

Example:

-- foo:bar~
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.fuzzy "bar")

gt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than range query.

gt e = excl e `to` star

Example:

-- foo:>5
-- foo:{5 TO *]
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.gt (Solr.int 5))

gte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than-or-equal-to range query.

gte e = incl e `to` star

Example:

-- foo:>=5
-- foo:[5 TO *]
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.gte (Solr.int 5))

lt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than range query.

 lt e = star `to` excl e

Example:

-- foo:<5
-- foo:[* TO 5}
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.lt (Solr.int 5))

lte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than-or-equal-to range query.

lte e = star `to` incl e

Example:

-- foo:<=5
-- foo:[* TO 5]
query :: Solr.SolrQuery 'False
query = Solr.field "foo" (Solr.lte (Solr.int 5))

Range query helpers

data Boundary a

An inclusive or exclusive expression for use in a range query, built with either incl, excl, or star.

The constructors are exported for use in interpreters.

Constructors

Inclusive a

Exclusive a

Star

incl :: SolrExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as inclusive, for use in a range query.

excl :: SolrExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as exclusive, for use in a range query.

star :: SolrExprSYM expr => Boundary (expr a)

'*' operator, signifying the minimum or maximun bound of a range. A [* TO *] query is allowed, but will require a type annotation.