Solr.Query

Description

This is the simplest interpretation of the Solr query language as a lazy Text.

Not all type-correct expressions using the Solr DSL result in well-formed queries. For example,

>>> let query = (("foo" =: word "bar") ^=: 1.0) ^=: 2.0 :: QuerySYM expr query => query expr
>>> compile [] (query :: Query Expr)
"q=foo:bar^=1.0^=2.0"

For this reason, you may want to first interpret a query using Solr.Query.Initial, manually fix up the AST (perhaps with factor), and then reinterpret it as the lazy Text version using reinterpret:

>>> import Solr.Query.Initial (factor, reinterpret)
>>> compile [] (reinterpret (factor query) :: Query Expr)
"q=foo:bar^=2.0"

Query

data Query expr

A Solr query.

compile :: [Param Query] -> Query expr -> Text

Compile a Query to a lazy Text.

Note that the DSL admits many ways to create an invalid Solr query (e.g. multiple negs); that is, if it compiles, it doesn't necessarily work.

Examples

>>> let query = "foo" =: phrase ["bar", "baz"] ~: 5 &&: defaultField (regex "wh?t") :: Query Expr
>>> compile [paramDefaultField "body"] query
"q={!df=body}(foo:\"bar baz\"~5 AND /wh?t/)"

Re-exports

module Solr.Expr

module Solr.Expr.Class

module Solr.Query.Class

module Solr.Param