Solr.Query

Description

Solr query construction and compilation.

Query type

data SolrQuery expr

A Solr query.

data SolrFilterQuery expr

A Solr filter query. This is like SolrQuery, but with different local parameters available. All functions polymorphic over SolrQuerySYM will work with both.

Query construction

defaultField :: SolrQuerySYM expr query => expr a -> query expr

A default field query.

>>> defaultField (word "foo") :: SolrQuery SolrExpr
q=foo

(=:) :: SolrQuerySYM expr query => Text -> expr a -> query expr

A field query.

>>> "foo" =: word "bar" :: SolrQuery SolrExpr
q=foo:bar

field :: SolrQuerySYM expr query => Text -> expr a -> query expr

Named version of (=:).

(&&:) :: SolrQuerySYM expr query => query expr -> query expr -> query expr

An AND query.

>>> "foo" =: word "bar" &&: "baz" =: word "qux" :: SolrQuery SolrExpr
q=(foo:bar AND baz:qux)

and :: SolrQuerySYM expr query => query expr -> query expr -> query expr infixr 3

Named version of (&&:).

(||:) :: SolrQuerySYM expr query => query expr -> query expr -> query expr

An OR query.

>>> "foo" =: word "bar" ||: "baz" =: word "qux" :: SolrQuery SolrExpr
q=(foo:bar OR baz:qux)

or :: SolrQuerySYM expr query => query expr -> query expr -> query expr infixr 2

Named version of (||:).

(-:) :: SolrQuerySYM expr query => query expr -> query expr -> query expr

A NOT, '!', or '-' query.

>>> "foo" =: word "bar" -: "baz" =: word "qux" :: SolrQuery SolrExpr
q=(foo:bar NOT baz:qux)

not :: SolrQuerySYM expr query => query expr -> query expr -> query expr infixr 1

Named version of (-:).

(^=:) :: SolrQuerySYM expr query => query expr -> Float -> query expr

The '^=' constant score operator.

This is given right-fixity to reject queries like q ^= 1 ^= 2.

>>> "foo" =: word "bar" ^=: 3.5 :: SolrQuery SolrExpr
q=foo:bar^=3.5

score :: SolrQuerySYM expr query => query expr -> Float -> query expr infixr 4

Named version of (^=:).

neg :: SolrQuerySYM expr query => query expr -> query expr

Negate a query.

>>> neg ("foo" =: word "bar") :: SolrQuery SolrExpr
q=-foo:bar

params :: SolrQuerySYM expr query => [Param query] -> query expr -> query expr

Add local parameters to a query.

>>> params [paramDefaultField .= "foo"] (defaultField (word "bar")) :: SolrQuery SolrExpr
q={!df=foo}bar

Expression type

data SolrExpr t

A Solr expression.

Expression construction

num :: SolrExprSYM expr => Float -> expr TNum

A num expression.

>>> "foo" =: num 5 :: SolrQuery SolrExpr
q=foo:5.0

true :: SolrExprSYM expr => expr TBool

A true expression.

>>> "foo" =: true :: SolrQuery SolrExpr
q=foo:true

false :: SolrExprSYM expr => expr TBool

A false expression.

>>> "foo" =: false :: SolrQuery SolrExpr
q=foo:false

word :: SolrExprSYM expr => Text -> expr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for SolrExpr TWord, but usually an explicit type signature will be required (at the interpretation site or earlier).

>>> "foo" =: word "bar" :: SolrQuery SolrExpr
q=foo:bar

wild :: SolrExprSYM expr => Text -> expr TWild

A single word that may contain wildcard characters ('?' and '*'), although the meaning of consecutive '*'s is probably ill-defined. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

>>> "foo" =: wild "b?r" :: SolrQuery SolrExpr
q=foo:b?r

regex :: SolrExprSYM expr => Text -> expr TRegex

A regular expression, whose syntax is described by http://lucene.apache.org/core/5_5_0/core/org/apache/lucene/util/automaton/RegExp.html?is-external=true.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

>>> "foo" =: regex "[mb]oat" :: SolrQuery SolrExpr
q=foo:/[mb]oat/

phrase :: SolrExprSYM expr => [expr TWord] -> expr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

>>> "foo" =: phrase ["bar", "baz"] :: SolrQuery SolrExpr
q=foo:"bar baz"

(~:) :: (SolrExprSYM expr, FuzzableType a) => expr a -> Int -> expr TFuzzed

The '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

>>> "foo" =: word "bar" ~: 1 :: SolrQuery SolrExpr
q=foo:bar~1

>>> "foo" =: phrase ["bar", "baz", "qux"] ~: 10 :: SolrQuery SolrExpr
q=foo:"bar baz qux"~10

fuzz :: (SolrExprSYM expr, FuzzableType a) => expr a -> Int -> expr TFuzzed

Named version of (~:).

fuzzy :: SolrExprSYM expr => expr TWord -> expr TFuzzed

Short-hand for fuzzing a word by 2. This is the default behavior of a Solr '~' operator without an integer added.

fuzzy e = e ~: 2

>>> "foo" =: fuzzy "bar" :: SolrQuery SolrExpr
q=foo:bar~2

to :: (SolrExprSYM expr, PrimType a) => Boundary (expr a) -> Boundary (expr a) -> expr TRange

A range expression.

>>> "foo" =: incl (num 5) `to` excl (num 10) :: SolrQuery SolrExpr
q=foo:[5.0 TO 10.0}

>>> "foo" =: excl (word "bar") `to` star :: SolrQuery SolrExpr
q=foo:{bar TO *]

- Note the explicit type signature required for [* TO *] queries >>> "foo" =: star to (star :: Boundary (SolrExpr 'TNum)) :: SolrQuery SolrExpr q=foo:[* TO *]

incl :: SolrExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as inclusive, for use in a range query.

excl :: SolrExprSYM expr => expr a -> Boundary (expr a)

Mark an expression as exclusive, for use in a range query.

star :: SolrExprSYM expr => Boundary (expr a)

'*' operator, signifying the minimum or maximun bound of a range. A [* TO *] query will require a type annotation.

gt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than range query.

gt e = excl e `to` star

>>> "foo" =: gt (num 5) :: SolrQuery SolrExpr
q=foo:{5.0 TO *]

gte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than-or-equal-to range query.

gte e = incl e `to` star

>>> "foo" =: gte (num 5) :: SolrQuery SolrExpr
q=foo:[5.0 TO *]

lt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than range query.

 lt e = star `to` excl e

>>> "foo" =: lt (num 5) :: SolrQuery SolrExpr
q=foo:[* TO 5.0}

lte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than-or-equal-to range query.

lte e = star `to` incl e

>>> "foo" =: lte (num 5) :: SolrQuery SolrExpr
q=foo:[* TO 5.0]

(^:) :: (SolrExprSYM expr, BoostableType a) => expr a -> Float -> expr TBoosted

The '^' operator, which boosts its argument.

>>> "foo" =: word "bar" ^: 3.5 :: SolrQuery SolrExpr
q=foo:bar^3.5

>>> "foo" =: phrase ["bar", "baz"] ^: 3.5 :: SolrQuery SolrExpr
q=foo:"bar baz"^3.5

boost :: (SolrExprSYM expr, BoostableType a) => expr a -> Float -> expr TBoosted

Named version of (^:).

Local parameters

data Param query where

A parameter is built from a key and a value, whose type depends on the key.

Constructors

Param :: ParamKey query val -> val -> Param query

(.=) :: ParamKey query val -> val -> Param query

Infix constructor for Param.

class HasParamDefaultField query where

The class of queries that support the 'df' local parameter.

Methods

paramDefaultField :: ParamKey query Text

The 'df' local parameter.

>>> params [paramDefaultField .= "foo"] (defaultField (word "bar")) :: SolrQuery SolrExpr
q={!df=foo}bar

class HasParamOp query where

The class of queries that support the 'op' local parameter.

Methods

paramOp :: ParamKey query Text

The 'op' local parameter.

Stringly typed to avoid a clunky sum type like

data Operator = And | Or | ...

which seems to have little value in cases like this. Instead, just pass "AND", "OR", ...

>>> params [paramOp .= "AND"] (defaultField (word "foo") <> defaultField (word "bar")) :: SolrQuery SolrExpr
q={!q.op=AND}foo bar

class HasParamCache query where

The class of queries that support the 'cache' local parameter.

Methods

paramCache :: ParamKey query Bool

The 'cache' local parameter.

>>> params [paramCache .= False] ("foo" =: word "bar") :: SolrFilterQuery SolrExpr
fq={!cache=false}foo:bar

class HasParamCost query where

The class of queries that support the 'cost' local parameter.

Methods

paramCost :: ParamKey query Int

The 'cost' local parameter.

>>> params [paramCost .= 5] ("foo" =: word "bar") :: SolrFilterQuery SolrExpr
fq={!cost=5}foo:bar

Query compilation

compileSolrQuery :: SolrQuery expr -> ByteString

Compile a SolrQuery to a lazy ByteString.

Note that the DSL admits many ways to create an invalid Solr query; that is, if it compiles, it doesn't necessarily work. For example, multiple negs on a query, multiple params, etc.

>>> let ps = [paramDefaultField .= "body"]
>>> let q = "foo" =: phrase ["bar", "baz"] ~: 5 &&: defaultField (regex "wh?t") :: SolrQuery SolrExpr
>>> compileSolrQuery (params ps q)
"q={!df=body}(foo:\"bar baz\"~5 AND /wh?t/)"

compileSolrFilterQuery :: SolrFilterQuery expr -> ByteString

Compile a SolrFilterQuery to a lazy ByteString.