Solr.Query

Description

Solr query construction and compilation. You may prefer to import Solr.Qualified.Query instead, which does not export any operators.

Query type

data SolrQuery isNeg hasParams

A Solr query.

The two boolean phantom types track whether or not this query has been negated, and whether or not this query has local parameters.

While this approach allows fewer bad queries to typecheck, it is not extensible, leaks abstraction, makes documentation more difficult to read, and basically suffers from type-level boolean blindness (a "Could not match True with False" type error is not very helpful). So, this might change eventually.

data SolrFilterQuery isNeg hasParams

A Solr filter query. This is like SolrQuery, but with different local parameters available. All functions polymorphic over SolrQuerySYM will work with both.

Query construction

For simplicity, the type signatures in the examples below monomorphise the functions to use SolrQuery (and therefore SolrExpr, due to the functional dependency).

defaultField :: SolrQuerySYM expr query => expr a -> query False False

A default field query.

Example:

-- foo
query :: SolrQuery 'False 'False
query = defaultField (word "foo")

(=:) :: SolrQuerySYM expr query => Text -> expr a -> query False False

A field query.

Example:

-- foo:bar
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"

(&&:) :: SolrQuerySYM expr query => query False False -> query False False -> query False False

An AND query.

Example:

-- foo:bar AND baz:qux
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"
    &&: "baz" =: word "qux"

(||:) :: SolrQuerySYM expr query => query False False -> query False False -> query False False

An OR query.

Example:

-- foo:bar OR baz:qux
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"
    ||: "baz" =: word "qux"

(-:) :: SolrQuerySYM expr query => query False False -> query False False -> query False False

A NOT, '!', or '-' query.

Example:

-- foo:bar NOT baz:qux
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"
     -: "baz" =: word "qux"

(^=:) :: SolrQuerySYM expr query => query False False -> Float -> query False False

The '^=' constant score operator.

This is given right-fixity to reject queries like q ^= 1 ^= 2, which may very well be a valid Solr query (I haven't tested), but are nonetheless nonsense.

Example:

-- (foo:bar)^=3.5
query :: SolrQuery 'False 'False
query = "foo" =: word "bar" ^=: 3.5

neg :: SolrQuerySYM expr query => query False False -> query True False

Negate a top-level query.

Only a top level query can be negated up to one time, hence the boolean tag that tracks whether or not a query has been negated.

Example:

-- -(foo:bar)
query :: SolrQuery 'True 'False
query = neg ("foo" =: word "bar")

params :: SolrQuerySYM expr query => [Param query] -> query isNeg False -> query isNeg True

Add local parameters to a query.

Only a top level query can have up to one set of local parameters, hence the boolean tag that tracks whether or not they've been added.

Example:

-- {!df=foo}bar
query :: SolrQuery 'False 'True
query = params [paramDefaultField .= "foo"] (defaultField (word "bar"))

Expression type

data SolrExpr t

A Solr expression.

Expression construction

For simplicity, the type signatures in the examples below monomorphise the functions to use SolrQuery (and therefore SolrExpr, due to the functional dependency).

num :: SolrExprSYM expr => Float -> expr TNum

A num expression.

Example:

-- foo:5
query :: SolrQuery 'False 'False
query = "foo" =: num 5

true :: SolrExprSYM expr => expr TBool

A true expression.

Example:

-- foo:true
query :: SolrQuery 'False 'False
query = "foo" =: true

false :: SolrExprSYM expr => expr TBool

A false expression.

Example:

-- foo:false
query :: SolrQuery 'False 'False
query = "foo" =: false

word :: SolrExprSYM expr => Text -> expr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for SolrExpr TWord, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:bar
query :: SolrQuery 'False 'False
query = "foo" =: word "bar"

-- foo:bar
query :: SolrQuery 'False 'False
query = "foo" =: ("bar" :: SolrExpr TWord)

wild :: SolrExprSYM expr => Text -> expr TWild

A single word that may contain wildcard characters ('?' and '*'), although the meaning of consecutive '*'s is probably ill-defined. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

Example:

-- foo:b?r
query :: SolrQuery 'False 'False
query = "foo" =: wild "b?r"

regex :: SolrExprSYM expr => Text -> expr TRegex

A regular expression, whose syntax is described by http://lucene.apache.org/core/5_5_0/core/org/apache/lucene/util/automaton/RegExp.html?is-external=true.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

-- foo:/[mb]oat/
query :: SolrQuery 'False 'False
query = "foo" =: regex "[mb]oat"

phrase :: SolrExprSYM expr => [expr TWord] -> expr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

Note that sometimes you may use the IsList instance for SolrExpr TPhrase, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:"bar baz"
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", "baz"] -- ok

-- foo:"bar b?z" (an invalid Solr query)
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", wild "b?z"] -- type error

-- foo:"bar b?z" (an invalid Solr query)
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", "b?z"] -- breaks word contract

Or, with OverloadedLists:

-- foo:"bar baz"
query :: SolrQuery 'False 'False
query = "foo" =: (["bar", "baz"] :: SolrExpr TPhrase)

(~:) :: (SolrExprSYM expr, FuzzableType a) => expr a -> Int -> expr TFuzzed

The '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

Example:

-- foo:bar~1
query :: SolrQuery 'False 'False
query = "foo" =: word "bar" ~: 1

-- foo:"bar baz qux"~10
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", "baz", "qux"] ~: 10

fuzzy :: SolrExprSYM expr => expr TWord -> expr TFuzzed

Short-hand for fuzzing a word by 2. This is the default behavior of a Solr '~' operator without an integer added.

fuzzy e = fuzz e 2

Example:

-- foo:bar~
query :: SolrQuery 'False 'False
query = "foo" =: fuzzy "bar"

to :: (SolrExprSYM expr, PrimType a) => Boundary (expr a) -> Boundary (expr a) -> expr TRange

A range expression.

Example:

-- foo:[5 TO 10}
query :: SolrQuery 'False 'False
query = "foo" =: incl (num 5) `to` excl (num 10)

gt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than range query.

gt e = excl e `to` star

Example:

-- foo:>5
-- foo:{5 TO *]
query :: SolrQuery 'False 'False
query = "foo" =: gt (num 5)

gte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than-or-equal-to range query.

gte e = incl e `to` star

Example:

-- foo:>=5
-- foo:[5 TO *]
query :: SolrQuery 'False 'False
query = "foo" =: gte (num 5)

lt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than range query.

 lt e = star `to` excl e

Example:

-- foo:<5
-- foo:[* TO 5}
query :: SolrQuery 'False 'False
query = "foo" =: lt (num 5)

lte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than-or-equal-to range query.

lte e = star `to` incl e

Example:

-- foo:<=5
-- foo:[* TO 5]
query :: SolrQuery 'False 'False
query = "foo" =: lte (num 5)

(^:) :: (SolrExprSYM expr, BoostableType a) => expr a -> Float -> expr TBoosted

The '^' operator, which boosts its argument.

Example:

-- foo:bar^3.5
query :: SolrQuery 'False 'False
query = "foo" =: word "bar" ^: 3.5

-- foo:"bar baz"^3.5
query :: SolrQuery 'False 'False
query = "foo" =: phrase ["bar", "baz"] ^: 3.5

Local parameters

data Param query where

A parameter is built from a key and a value, whose type depends on the key.

Constructors

Param :: ParamKey query val -> val -> Param query

(.=) :: ParamKey query val -> val -> Param query

Infix constructor for Param.

class ParamDefaultField query where

The class of queries that support the 'df' local parameter.

Methods

paramDefaultField :: ParamKey query Text

The 'df' local parameter.

Example:

-- {!df=foo}bar
query :: SolrQuery 'False 'True
query = params [paramDefaultField .= "foo"] (defaultField (word "bar"))

class ParamOp query where

The class of queries that support the 'op' local parameter.

Methods

paramOp :: ParamKey query Text

The 'op' local parameter.

Stringly typed to avoid a clunky sum type like

data Val = And | Or | ...

which seems to have little value in cases like this. Instead, just pass "AND", "OR", ...

Example:

-- {!q.op=AND}foo bar
query :: SolrQuery 'False 'True
query = params [paramOp .= "AND"] (defaultField (word "foo") <> defaultField (word "bar"))

paramCache :: ParamKey SolrFilterQuery Bool

The 'cache' local parameter.

Example:

-- {!cache=false}foo:bar
query :: SolrFilterQuery 'False 'True
query = params [paramCache .= False] ("foo" =: word "bar")

Query compilation

compileSolrQuery :: SolrQuery isNeg hasParams -> ByteString

Compile a SolrQuery to a lazy ByteString.

Example:

λ let ps = [paramDefaultField .= "body"]
λ let q = "foo" =: phrase ["bar", "baz"] ~: 5 &&: defaultField (regex "wh?at")
λ compileSolrQuery (params ps q)
"{!df=body}(foo:\"bar baz\"~5 AND /wh?t/)"

compileSolrFilterQuery :: SolrFilterQuery isNeg hasParams -> ByteString

Compile a SolrFilterQuery to a lazy ByteString.