Solr.Query

Description

Solr query construction and compilation. You may prefer to import Solr.Qualified.Query instead, which does not export any operators.

Query type

data SolrQuery :: Bool -> *

A Solr query. The boolean phantom type tracks whether or not this query has local params or not.

You may ignore the data LocalParams SolrQuery instance below; the data constructor SolrQueryParams is not exported, but shows up in the Haddocks anyway.

Query construction

For simplicity, the type signatures in the examples below monomorphise the functions to use SolrQuery (and therefore SolrExpr, due to the functional dependency).

defaultField :: SolrQuerySYM expr query => expr a -> query False

A default field query.

Example:

-- foo
query :: SolrQuery 'False
query = defaultField (word "foo")

(=:) :: SolrQuerySYM expr query => Text -> expr a -> query False

A field query.

Example:

-- foo:bar
query :: SolrQuery 'False
query = "foo" =: word "bar"

(&&:) :: SolrQuerySYM expr query => query False -> query False -> query False

An AND query.

Example:

-- foo:bar AND baz:qux
query :: SolrQuery 'False
query = "foo" =: word "bar"
    &&: "baz" =: word "qux"

(||:) :: SolrQuerySYM expr query => query False -> query False -> query False

An OR query.

Example:

-- foo:bar OR baz:qux
query :: SolrQuery 'False
query = "foo" =: word "bar"
    ||: "baz" =: word "qux"

(-:) :: SolrQuerySYM expr query => query False -> query False -> query False

A NOT, '!', or '-' query.

Example:

-- foo:bar NOT baz:qux
query :: SolrQuery 'False
query = "foo" =: word "bar"
     -: "baz" =: word "qux"

(^=:) :: SolrQuerySYM expr query => query False -> Float -> query False

The '^=' constant score operator.

This is given right-fixity to reject queries like q ^= 1 ^= 2, which may very well be a valid Solr query (I haven't tested), but are nonetheless nonsense.

Example:

-- (foo:bar)^=3.5
query :: SolrQuery 'False
query = "foo" =: word "bar" ^=: 3.5

localParams :: SolrQuerySYM expr query => LocalParams query -> query False -> query True

Add local parameters to a query. A query can only have one set of local parameters, hence the boolean tag that tracks whether or not they've been added.

Example:

-- {!df=foo}bar
query :: SolrQuery 'True
query = localParams (paramDefaultField "foo") (defaultField (word "bar"))

Expression type

data SolrExpr t

A Solr expression.

Expression construction

For simplicity, the type signatures in the examples below monomorphise the functions to use SolrQuery (and therefore SolrExpr, due to the functional dependency).

int :: SolrExprSYM expr => Int -> expr TInt

An int expression.

Note that sometimes you may use the Num instance for SolrExpr TInt, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:5
query :: SolrQuery 'False
query = "foo" =: int 5

true :: SolrExprSYM expr => expr TBool

A true expression.

Example:

-- foo:true
query :: SolrQuery 'False
query = "foo" =: true

false :: SolrExprSYM expr => expr TBool

A false expression.

Example:

-- foo:false
query :: SolrQuery 'False
query = "foo" =: false

word :: SolrExprSYM expr => Text -> expr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for SolrExpr TWord, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:bar
query :: SolrQuery 'False
query = "foo" =: word "bar"

-- foo:bar
query :: SolrQuery 'False
query = "foo" =: ("bar" :: SolrExpr TWord)

wild :: SolrExprSYM expr => Text -> expr TWild

A single word that may contain wildcard characters ('?' and '*'), although the meaning of consecutive '*'s is probably ill-defined. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

Example:

-- foo:b?r
query :: SolrQuery 'False
query = "foo" =: wild "b?r"

regex :: SolrExprSYM expr => Text -> expr TRegex

A regular expression, whose syntax is described by http://lucene.apache.org/core/5_5_0/core/org/apache/lucene/util/automaton/RegExp.html?is-external=true.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

-- foo:/[mb]oat/
query :: SolrQuery 'False
query = "foo" =: regex "[mb]oat"

phrase :: SolrExprSYM expr => [expr TWord] -> expr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

Note that sometimes you may use the IsList instance for SolrExpr TPhrase, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:"bar baz"
query :: SolrQuery 'False
query = "foo" =: phrase ["bar", "baz"] -- ok

-- foo:"bar b?z" (an invalid Solr query)
query :: SolrQuery 'False
query = "foo" =: phrase ["bar", wild "b?z"] -- type error

-- foo:"bar b?z" (an invalid Solr query)
query :: SolrQuery 'False
query = "foo" =: phrase ["bar", "b?z"] -- breaks word contract

Or, with OverloadedLists:

-- foo:"bar baz"
query :: SolrQuery 'False
query = "foo" =: (["bar", "baz"] :: SolrExpr TPhrase)

(~:) :: (SolrExprSYM expr, FuzzableType a) => expr a -> Int -> expr (TFuzzed a)

The '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

This will have one of the following two types:

(~:) :: SolrExpr TWord   -> Int -> SolrExpr TFuzzyWord   -- Int must be 0, 1, or 2
(~:) :: SolrExpr TPhrase -> Int -> SolrExpr TFuzzyPhrase -- Int must be positive

Example:

-- foo:bar~1
query :: SolrQuery 'False
query = "foo" =: word "bar" ~: 1

-- foo:"bar baz qux"~10
query :: SolrQuery 'False
query = "foo" =: phrase ["bar", "baz", "qux"] ~: 10

fuzzy :: SolrExprSYM expr => expr TWord -> expr TFuzzyWord

Short-hand for fuzzing a word by 2. This is the default behavior of a Solr '~' operator without an integer added.

fuzzy e = fuzz e 2

Example:

-- foo:bar~
query :: SolrQuery 'False
query = "foo" =: fuzzy "bar"

to :: (SolrExprSYM expr, PrimType a) => Boundary (expr a) -> Boundary (expr a) -> expr TRange

A range expression.

This will have one of the following two types:

to :: Boundary (SolrExpr TWord) -> Boundary (SolrExpr TWord) -> SolrExpr TRange
to :: Boundary (SolrExpr TInt)  -> Boundary (SolrExpr TInt)  -> SolrExpr TRange

Example:

-- foo:[5 TO 10}
query :: SolrQuery 'False
query = "foo" =: incl (int 5) `to` excl (int 10)

gt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than range query.

gt e = excl e `to` star

Example:

-- foo:>5
-- foo:{5 TO *]
query :: SolrQuery 'False
query = "foo" =: gt (int 5)

gte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than-or-equal-to range query.

gte e = incl e `to` star

Example:

-- foo:>=5
-- foo:[5 TO *]
query :: SolrQuery 'False
query = "foo" =: gte (int 5)

lt :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than range query.

 lt e = star `to` excl e

Example:

-- foo:<5
-- foo:[* TO 5}
query :: SolrQuery 'False
query = "foo" =: lt (int 5)

lte :: (SolrExprSYM expr, PrimType a) => expr a -> expr TRange

Short-hand for a less-than-or-equal-to range query.

lte e = star `to` incl e

Example:

-- foo:<=5
-- foo:[* TO 5]
query :: SolrQuery 'False
query = "foo" =: lte (int 5)

(^:) :: (SolrExprSYM expr, BoostableType a) => expr a -> Float -> expr (TBoosted a)

The '^' operator, which boosts its argument.

This will have one of the following two types:

(^:) :: SolrExpr TWord   -> Float -> SolrExpr TBoostedWord
(^:) :: SolrExpr TPhrase -> Float -> SolrExpr TBoostedPhrase

Example:

-- foo:bar^3.5
query :: SolrQuery 'False
query = "foo" =: word "bar" ^: 3.5

-- foo:"bar baz"^3.5
query :: SolrQuery 'False
query = "foo" =: phrase ["bar", "baz"] ^: 3.5

Local parameters

Combine LocalParams SolrQuery with (<>) and mark a query as having local parameters with localParams.

paramDefaultField :: Text -> LocalParams SolrQuery

Set the 'df' local parameter.

Example:

-- {!df=foo}bar
query :: SolrQuery 'True
query = localParams (paramDefaultField "foo") (defaultField (word "bar"))

Query compilation

compileSolrQuery :: SolrQuery a -> ByteString

Compile a SolrQuery to a lazy ByteString.

Example:

λ let params = paramDefaultField "body"
λ let query = "foo" =: phrase ["bar", "baz"] ~: 5 &&: defaultField (regex "wh?at")
λ compileSolrQuery (localParams params query)
"{!df=body}(foo:"bar baz"~5 AND wh?t)"