Solr.Query.Initial

Description

An initial encoding of a Solr query. This is an alternative interpretation of the Solr language that is more amenable to parsing from arbitrary user input.

Comes in typed and untyped variants; see Solr.Expr.Initial.Typed and Solr.Expr.Initial.Untyped.

Query type

data SolrQuery expr

A Solr query.

Constructors

forall a . QDefaultField (expr a)

forall a . QField Text (expr a)

QAnd (SolrQuery expr) (SolrQuery expr)

QOr (SolrQuery expr) (SolrQuery expr)

QNot (SolrQuery expr) (SolrQuery expr)

QScore (SolrQuery expr) Float

QNeg (SolrQuery expr)

QParams [Param SolrQuery] (SolrQuery expr)

Query construction

defaultField :: SolrQuerySYM expr query => expr a -> query expr

A default field query.

>>> defaultField (word "foo") :: SolrQuery SolrExpr
q=foo

(=:) :: SolrQuerySYM expr query => Text -> expr a -> query expr

A field query.

>>> "foo" =: word "bar" :: SolrQuery SolrExpr
q=foo:bar

field :: SolrQuerySYM expr query => Text -> expr a -> query expr

Named version of (=:).

(&&:) :: SolrQuerySYM expr query => query expr -> query expr -> query expr

An AND query.

>>> "foo" =: word "bar" &&: "baz" =: word "qux" :: SolrQuery SolrExpr
q=(foo:bar AND baz:qux)

and :: SolrQuerySYM expr query => query expr -> query expr -> query expr infixr 3

Named version of (&&:).

(||:) :: SolrQuerySYM expr query => query expr -> query expr -> query expr

An OR query.

>>> "foo" =: word "bar" ||: "baz" =: word "qux" :: SolrQuery SolrExpr
q=(foo:bar OR baz:qux)

or :: SolrQuerySYM expr query => query expr -> query expr -> query expr infixr 2

Named version of (||:).

(-:) :: SolrQuerySYM expr query => query expr -> query expr -> query expr

A NOT, '!', or '-' query.

>>> "foo" =: word "bar" -: "baz" =: word "qux" :: SolrQuery SolrExpr
q=(foo:bar NOT baz:qux)

not :: SolrQuerySYM expr query => query expr -> query expr -> query expr infixr 1

Named version of (-:).

(^=:) :: SolrQuerySYM expr query => query expr -> Float -> query expr

The '^=' constant score operator.

This is given right-fixity to reject queries like q ^= 1 ^= 2.

>>> "foo" =: word "bar" ^=: 3.5 :: SolrQuery SolrExpr
q=foo:bar^=3.5

score :: SolrQuerySYM expr query => query expr -> Float -> query expr infixr 4

Named version of (^=:).

neg :: SolrQuerySYM expr query => query expr -> query expr

Negate a query.

>>> neg ("foo" =: word "bar") :: SolrQuery SolrExpr
q=-foo:bar

params :: SolrQuerySYM expr query => [Param query] -> query expr -> query expr

Add local parameters to a query.

>>> params [paramDefaultField .= "foo"] (defaultField (word "bar")) :: SolrQuery SolrExpr
q={!df=foo}bar

Local parameters

data Param query where

A parameter is built from a key and a value, whose type depends on the key.

Constructors

Param :: ParamKey query val -> val -> Param query

(.=) :: ParamKey query val -> val -> Param query

Infix constructor for Param.

class HasParamDefaultField query where

The class of queries that support the 'df' local parameter.

Methods

paramDefaultField :: ParamKey query Text

The 'df' local parameter.

>>> params [paramDefaultField .= "foo"] (defaultField (word "bar")) :: SolrQuery SolrExpr
q={!df=foo}bar

class HasParamOp query where

The class of queries that support the 'op' local parameter.

Methods

paramOp :: ParamKey query Text

The 'op' local parameter.

Stringly typed to avoid a clunky sum type like

data Operator = And | Or | ...

which seems to have little value in cases like this. Instead, just pass "AND", "OR", ...

>>> params [paramOp .= "AND"] (defaultField (word "foo") <> defaultField (word "bar")) :: SolrQuery SolrExpr
q={!q.op=AND}foo bar

class HasParamCache query where

The class of queries that support the 'cache' local parameter.

Methods

paramCache :: ParamKey query Bool

The 'cache' local parameter.

>>> params [paramCache .= False] ("foo" =: word "bar") :: SolrFilterQuery SolrExpr
fq={!cache=false}foo:bar

class HasParamCost query where

The class of queries that support the 'cost' local parameter.

Methods

paramCost :: ParamKey query Int

The 'cost' local parameter.

>>> params [paramCost .= 5] ("foo" =: word "bar") :: SolrFilterQuery SolrExpr
fq={!cost=5}foo:bar

Type checking

typeCheckSolrQuery :: SolrQuery SolrExpr -> Maybe (SolrQuery SolrExpr)

Typecheck an untyped Solr query. Note the SolrExpr on the way in is not the same as the SolrExpr on the way out.