Solr.Query.Initial

Description

An initial encoding of a Solr query. This is an alternative interpretation of the Solr language that is more amenable to parsing from arbitrary user input and applying query transformations.

Query type

data SolrQuery expr

An initial encoding of a Solr query.

Constructors

forall a . QDefaultField (expr a)

forall a . QField Text (expr a)

QAnd (SolrQuery expr) (SolrQuery expr)

QOr (SolrQuery expr) (SolrQuery expr)

QNot (SolrQuery expr) (SolrQuery expr)

QScore (SolrQuery expr) Float

QNeg (SolrQuery expr)

QParams [Param SolrQuery] (SolrQuery expr)

QAppend (SolrQuery expr) (SolrQuery expr)

Type checking

typeCheckSolrQuery :: SolrQuery SolrExpr -> Maybe (SolrQuery SolrExpr)

Type check an untyped Solr query. Note the untyped SolrExpr on the way in is not the same as the typed SolrExpr on the way out.

Factorization

factorSolrQuery :: SolrQuery expr -> SolrQuery expr

Attempt to factor a Solr query into a canonical form that irons out invalid queries that are not caught by the type system (for example, double-negation, or multiple nested applications of params).

Check the source code for all transformations performed.

Reinterpretation

reinterpretSolrQuery :: forall expr query. (SolrQuerySYM expr query, HasParamDefaultField query, HasParamOp query, Semigroup (query expr)) => SolrQuery SolrExpr -> query expr

Reinterpret an initially-encoded SolrQuery to some other interpretation that supports all of 'SolrQuery'\'s params.

This may be useful for reinterpreting a SolrQuery as a lazy ByteString after it's been type checked and factored with the machinery in this module.

Re-exports

module Solr.Internal.Class.Query

module Solr.Query.Param