Solr.Query.Initial

Description

An initial encoding of a Solr query. This is an alternative interpretation of the Solr language that is more amenable to parsing from arbitrary user input.

Query type

data SolrQueryI where

A Solr query.

Constructors

QDefaultField :: SolrExprI a -> SolrQueryI

QField :: Text -> SolrExprI a -> SolrQueryI

QAnd :: SolrQueryI -> SolrQueryI -> SolrQueryI

QOr :: SolrQueryI -> SolrQueryI -> SolrQueryI

QNot :: SolrQueryI -> SolrQueryI -> SolrQueryI

QScore :: SolrQueryI -> Float -> SolrQueryI

QNeg :: SolrQueryI -> SolrQueryI

QParams :: [Param SolrQueryI] -> SolrQueryI -> SolrQueryI

Expression type

data SolrExprI ty where

A Solr expression.

Constructors

ENum :: Float -> SolrExprI TNum

ETrue :: SolrExprI TBool

EFalse :: SolrExprI TBool

EWord :: Text -> SolrExprI TWord

EWild :: Text -> SolrExprI TWild

ERegex :: Text -> SolrExprI TRegex

EPhrase :: [SolrExprI TWord] -> SolrExprI TPhrase

EFuzz :: FuzzableType a => SolrExprI a -> Int -> SolrExprI TFuzzed

ETo :: PrimType a => Boundary (SolrExprI a) -> Boundary (SolrExprI a) -> SolrExprI TRange

EBoost :: BoostableType a => SolrExprI a -> Float -> SolrExprI TBoosted

Query compilation

solrQueryI :: SolrQueryI -> SolrQueryI

Select the SolrQueryI interpreter.

solrQueryI = id

Untyped ADTs

data USolrQueryI

An untyped Solr query.

Constructors

UQDefaultField USolrExprI

UQField Text USolrExprI

UQAnd USolrQueryI USolrQueryI

UQOr USolrQueryI USolrQueryI

UQNot USolrQueryI USolrQueryI

UQScore USolrQueryI Float

UQNeg USolrQueryI

data USolrExprI

An untyped Solr expression.

Constructors

UENum Float

UETrue

UEFalse

UEWord Text

UEWild Text

UERegex Text

UEPhrase [USolrExprI]

UEFuzz USolrExprI Int

UETo (Boundary USolrExprI) (Boundary USolrExprI)

UEBoost USolrExprI Float

Type checking

typeCheckSolrQuery :: USolrQueryI -> r -> (SolrQueryI -> r) -> r